Improve coding style.
